+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `file-integrity` will be documented in this file.
+
+## 0.0.1 - Initial release
+
+- Initial release of the `file-integrity` package.